+++ src/modules/ColorTemperature/index.js
